Update BuySellSignalbyEMA
*** Use it at your own risk 

This indicator has 3 indicators and gives a buy/sell signal depending on the EMA.

3 indicator
 - Ichimoku
- Bollinger Band
- EMA
			
It also shows the value on the candle as a warning in the Buy and Sell signals. This value is calculated with a special formula I have created.

On the other hand Bollinger band and ichimoku cloud are also included to check the accuracy of the signals.

For this indicator, I recommend a minimum of 4 hours chart.